[unified-server] Move ServerDeploymentConfiguration out of common types
It's only used in unified-server. There's no need for it to be in common
types. This is a precursor to removing the type entirely.
