+++ packages/types/src/deployment-configuration.ts
@@ -36,20 +36,3 @@ export interface DomainConfiguration {
    */
   disallowPublicPublishing?: boolean;
 }
-
-export type ServerDeploymentConfiguration = {
-  BACKEND_API_ENDPOINT?: string;
-  /**
-   * The URL of the deployed server.
-   */
-  SERVER_URL?: string;
-  /**
-   * The Drive Id of a folder containing featured gallery items
-   */
-  GOOGLE_DRIVE_FEATURED_GALLERY_FOLDER_ID?: string;
-  /**
-   * The list of all MCP Servers allowed by the mcp proxy. Glob patterns
-   * accepted.
-   */
-  MCP_SERVER_ALLOW_LIST?: string[];
-};
