+++ packages/unified-server/src/server/provide-config.ts
@@ -8,13 +8,29 @@ import { readFile } from "node:fs/promises";
 
 import {
   type ClientDeploymentConfiguration,
-  type ServerDeploymentConfiguration,
   type DomainConfiguration,
 } from "@breadboard-ai/types/deployment-configuration.js";
 
 export type DeploymentConfiguration = {
   client: ClientDeploymentConfiguration;
   server: ServerDeploymentConfiguration;
+};
+
+export type ServerDeploymentConfiguration = {
+  BACKEND_API_ENDPOINT?: string;
+  /**
+   * The URL of the deployed server.
+   */
+  SERVER_URL?: string;
+  /**
+   * The Drive Id of a folder containing featured gallery items
+   */
+  GOOGLE_DRIVE_FEATURED_GALLERY_FOLDER_ID?: string;
+  /**
+   * The list of all MCP Servers allowed by the mcp proxy. Glob patterns
+   * accepted.
+   */
+  MCP_SERVER_ALLOW_LIST?: string[];
 };
 
 export async function getConfig(): Promise<DeploymentConfiguration> {
